Completed Spring Security with user registration and unique email id check along with validation details

+++ src/main/java/com/khozema/iniitian/controller/ProductController.java
@@ -29,7 +29,7 @@ public class ProductController {
 	}
 
 	@RequestMapping(value = "/{id}")
-	public @ResponseBody Product get(@PathVariable("id") int id) {
+	public @ResponseBody Product get(@PathVariable("id") Long id) {
 		return this.productService.get(id);
 	}
