+++ src/main/java/com/khozema/iniitian/dao/ProductDAO.java
@@ -6,12 +6,12 @@ import com.khozema.iniitian.entity.Category;
 import com.khozema.iniitian.entity.Product;
 
 public interface ProductDAO {
-	public Product get(int id);
+	public Product get(Long id);
 	public List<Product> getAll();
 	public Category getCategory(int id);
 	public List<Product> getProductsByCategory(int id);
 	public void add(Product c);
 	public void update(Product c);
-	public void remove(int id);	
+	public void remove(Long id);	
 
 }
